Create Word Search
The word_search function takes a list of document names and a keyword, then returns a list of index values of where in the original list that keyword pops up. Furthermore, it ignores differences in uppercase vs. lowercase and ignores any punctuation marks. 

The multi_word_search function fulfills the same purpose as the word_search function, except this takes in a list of keywords and returns a dictionary, with each key as the keyword and the value is a list of indices where that keyword is found.